+++ zarrs_codec/src/lib.rs
@@ -22,7 +22,7 @@ pub use array_bytes_fixed_disjoint_view::{
 
 mod codec_traits;
 pub use codec_traits::array::ArrayCodecTraits;
-pub use codec_traits::array_partial_sync::{ArrayPartialDecoderTraits, ArrayPartialEncoderTraits};
+pub use codec_traits::array_partial::{ArrayPartialDecoderTraits, ArrayPartialEncoderTraits};
 pub use codec_traits::array_to_array::{
     ArrayToArrayCodecSubchunkingIdentityTraits, ArrayToArrayCodecSubchunkingTraits,
     ArrayToArrayCodecTraits, UnboundArrayToArrayCodecTraits,
@@ -31,16 +31,16 @@ pub use codec_traits::array_to_bytes::{
     ArrayToBytesCodecNoSubchunkingTraits, ArrayToBytesCodecSubchunkingTraits,
     ArrayToBytesCodecTraits, UnboundArrayToBytesCodecTraits,
 };
-pub use codec_traits::bytes_partial_sync::{BytesPartialDecoderTraits, BytesPartialEncoderTraits};
+pub use codec_traits::bytes_partial::{BytesPartialDecoderTraits, BytesPartialEncoderTraits};
 pub use codec_traits::bytes_to_bytes::BytesToBytesCodecTraits;
 pub use codec_traits::{CodecTraits, CodecTraitsV2, CodecTraitsV3};
 
 #[cfg(feature = "async")]
-pub use codec_traits::array_partial_async::{
+pub use codec_traits::array_partial::{
     AsyncArrayPartialDecoderTraits, AsyncArrayPartialEncoderTraits,
 };
 #[cfg(feature = "async")]
-pub use codec_traits::bytes_partial_async::{
+pub use codec_traits::bytes_partial::{
     AsyncBytesPartialDecoderTraits, AsyncBytesPartialEncoderTraits,
 };